moved set_config.py to scripts
