+++ scripts/__init__.py
@@ -1,0 +1,8 @@
+# pylint: disable=unused-variable
+''' The Unified Workflow Tools package. '''
+
+__docformat__ = "NumPy/SciPy"
+
+import os
+
+uwtools_directory = os.path.dirname(__file__)
